Refact: extract shared pebble-pile module for the 3 take-away siblings
three-more, waning-stones and doubling-reduction each carried a
byte-for-byte identical Board type, cap helper, StonePile component,
BoardClient wrapper and getPlayerStepDescription. Move them to
single-pile-removal/pebble-pile.tsx; each game keeps only its own
strategy math, maxTake rule, start boards and rules text. cap is
re-exported from each game so spec imports stay unchanged.
