+++ src/components/games/single-pile-removal/waning-stones/waning-stones.tsx
@@ -1,19 +1,14 @@
-import {
-  strategyGameFactory, type Events, type StrategyArgs, type BoardClientProps, GameBoard, useHoverPreview
-} from '../../../strategy-game-factory';
+import { strategyGameFactory, type Events, type StrategyArgs } from '../../../strategy-game-factory';
 import { range, random, sample, minBy } from 'lodash';
-import { useTranslation } from '../../../../language';
+import { type Board, cap, BoardClient, getPlayerStepDescription } from '../pebble-pile';
 
-// `stones` is the number of pebbles left in the pile. `maxTake` is the most a
-// player may take this turn: on the opening move it is ⌊stones/2⌋ (baked into the
-// start board); after every move it becomes the amount just taken.
-type Board = { stones: number; maxTake: number }
+export { cap };
+
+// `maxTake` rule: on the opening move it is ⌊stones/2⌋ (baked into the start
+// board); after every move it becomes the amount just taken.
 
 // t(n): the largest power of 2 dividing n (its lowest set bit).
 export const lowestPow2 = (n: number): number => n & -n;
-
-// The most a player may legally take this turn.
-export const cap = (board: Board): number => Math.min(board.maxTake, board.stones);
 
 // A position (n stones, m max-take) is losing for the mover exactly when m < t(n).
 // So taking k wins iff it clears the pile, or hands the opponent such a losing
@@ -38,66 +33,6 @@ export const chooseSmartTake = (board: Board): number => {
     const oppCap = Math.min(k, rem);
     return range(1, oppCap + 1).filter(j => isWinningTake(rem, j)).length;
   })!;
-};
-
-const StonePile = ({ board, disabled, onTake, moveCount }: {
-  board: Board; disabled: boolean; onTake: (count: number) => void; moveCount: number
-}) => {
-  const { value, hoverProps } = useHoverPreview<number>(moveCount);
-  const previewCount = value ?? 0;
-  const c = cap(board);
-
-  return (
-    // rotate(180deg) makes the pile fill from the bottom (incomplete row on top)
-    // while keeping left-to-right reading order, so pebble "1" is the top-left of
-    // the pile and the count grows down and to the right. Each pebble re-rotates
-    // 180° so its number stays upright. Pebbles are taken from the top, so DOM
-    // index i corresponds to taking stones - i.
-    <div className="flex flex-wrap justify-center gap-1.5 p-2" style={{ transform: 'rotate(180deg)' }}>
-      {range(board.stones).map(i => {
-        const takeCount = board.stones - i; // clicking a pebble takes it and everything above
-        const takeable = takeCount <= c;
-        // Only selectable pebbles drive the hover preview. Don't rely on the
-        // `disabled` attribute to suppress this — some browsers (e.g. Safari)
-        // still fire pointer events on disabled buttons.
-        const canSelect = !disabled && takeable;
-        return (
-          <button
-            key={i}
-            disabled={disabled || !takeable}
-            onClick={() => onTake(takeCount)}
-            {...(canSelect ? hoverProps(takeCount) : {})}
-            aria-label={`${takeCount}`}
-            className={`w-[11%] sm:w-[8%] aspect-square rounded-full bg-stone-500 shadow-md shadow-stone-700
-              flex items-center justify-center text-white font-semibold text-xs sm:text-sm
-              transition-opacity
-              ${canSelect && takeCount <= previewCount ? 'opacity-30' : ''}`}
-            style={{ transform: 'rotate(180deg)' }}
-          >
-            {/* Active pebbles show how many would be removed by clicking them. */}
-            {canSelect && <span>{takeCount}</span>}
-          </button>
-        );
-      })}
-    </div>
-  );
-};
-
-const BoardClient = ({ board, ctx, moves }: BoardClientProps<Board>) => {
-  const { t } = useTranslation();
-  return (
-    <GameBoard>
-      <p className="text-center text-2xl font-bold mb-2">
-        {t({ hu: 'Kavicsok', en: 'Pebbles' })}: {board.stones}
-      </p>
-      <StonePile
-        board={board}
-        disabled={!ctx.isClientMoveAllowed}
-        onTake={count => moves.take(board, count)}
-        moveCount={ctx.moveCount}
-      />
-    </GameBoard>
-  );
 };
 
 const moves = {
@@ -149,11 +84,6 @@ const rule = {
     many as the other took last time. Whoever takes the last pebble(s) wins.
   </>
 };
-
-const getPlayerStepDescription = ({ board }: { board: Board }) => ({
-  hu: `Kattints egy kavicsra: legfeljebb ${cap(board)} kavicsot vehetsz el.`,
-  en: `Click a pebble: you may take at most ${cap(board)} pebble(s).`
-});
 
 export const WaningStones = strategyGameFactory({
   presentation: {
